+++ packages/domain/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./result.js";
 export * from "./routes.js";
 export * from "./ownership.js";
+export * from "./resource-contract.js";
 export * from "./types.js";